django-app/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
cssselect2==0.7.0
dj-database-url==0.5.0
Django==5.0
django-cors-headers==4.3.1
django-heroku==0.3.1
djangorestframework==3.14.0
fonttools==4.47.0
freetype-py==2.3.0
gunicorn==21.2.0
html5lib==1.1
idna==3.6
install==1.3.5
lxml==5.0.0
oscrypto==1.3.0
packaging==23.2
pdfkit==1.0.0
pillow==10.2.0
psycopg2==2.9.9
pycairo==1.25.1
pycparser==2.21
pydyf==0.8.0
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==3.17.4
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
razorpay==1.4.1
reportlab==4.0.8
requests==2.31.0
rlPyCairo==0.3.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.9.0
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
utils==1.0.1
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.13
zopfli==0.2.3
